refactor: extract sdl3 frontend
